src/pp/edward/inferences/__init__.py
"""
"""
from __future__ import absolute_import
from __future__ import division
from __future__ import print_function

from pp.edward.inferences.bigan_inference import *
from pp.edward.inferences.conjugacy import *
from pp.edward.inferences.gan_inference import *
from pp.edward.inferences.gibbs import *
from pp.edward.inferences.hmc import *
from pp.edward.inferences.implicit_klqp import *
from pp.edward.inferences.inference import *
from pp.edward.inferences.klpq import *
from pp.edward.inferences.klqp import *
from pp.edward.inferences.laplace import *
from pp.edward.inferences.map import *
from pp.edward.inferences.metropolis_hastings import *
from pp.edward.inferences.monte_carlo import *
from pp.edward.inferences.sgld import *
from pp.edward.inferences.sghmc import *
from pp.edward.inferences.variational_inference import *
from pp.edward.inferences.wake_sleep import *
from pp.edward.inferences.wgan_inference import *

from tensorflow.python.util.all_util import remove_undocumented

_allowed_symbols = [
    'BiGANInference',
    'complete_conditional',
    'GANInference',
    'Gibbs',
    'HMC',
    'ImplicitKLqp',
    'Inference',
    'KLpq',
    'KLqp',
    'ReparameterizationKLqp',
    'ReparameterizationKLKLqp',
    'ReparameterizationEntropyKLqp',
    'ScoreKLqp',
    'ScoreKLKLqp',
    'ScoreEntropyKLqp',
    'ScoreRBKLqp',
    'Laplace',
    'MAP',
    'MetropolisHastings',
    'MonteCarlo',
    'SGLD',
    'SGHMC',
    'VariationalInference',
    'WakeSleep',
    'WGANInference',
]

remove_undocumented(__name__, allowed_exception_list=_allowed_symbols)
