requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
appnope==0.1.0            # via ipython
bleach==1.5.0             # via nbconvert, tensorflow-tensorboard
cycler==0.10.0            # via matplotlib
decorator==4.1.2          # via ipython, traitlets
# edward==1.3.4             # changed by keyonvafa from 1.3.3 to 1.3.4
entrypoints==0.2.3        # via nbconvert
flake8==3.4.1             # via pytest-flake8
html5lib==0.9999999       # via bleach, tensorflow-tensorboard
ipykernel==4.6.1          # via ipywidgets, jupyter, jupyter-console, notebook, qtconsole
ipython-genutils==0.2.0   # via nbformat, notebook, qtconsole, traitlets
ipython==6.1.0            # via ipykernel, ipywidgets, jupyter-console
ipywidgets==7.0.0         # via jupyter
jedi==0.10.2              # via ipython
jinja2==2.9.6             # via nbconvert, notebook
jsonschema==2.6.0         # via nbformat
jupyter-client==5.1.0     # via ipykernel, jupyter-console, notebook, qtconsole
jupyter-console==5.2.0    # via jupyter
jupyter-core==4.3.0       # via jupyter-client, nbconvert, nbformat, notebook, qtconsole
jupyter==1.0.0
markdown==2.6.9           # via tensorflow-tensorboard
markupsafe==1.0           # via jinja2
matplotlib==2.0.2
mccabe==0.6.1             # via flake8
mistune==0.7.4            # via nbconvert
nbconvert==5.3.1          # via jupyter, notebook
nbformat==4.4.0           # via ipywidgets, nbconvert, notebook
notebook==5.0.0           # via jupyter, widgetsnbextension
numpy==1.13.1
olefile==0.44             # via pillow
pandas==0.20.3
pandocfilters==1.4.2      # via nbconvert
pexpect==4.2.1            # via ipython
pickleshare==0.7.4        # via ipython
pillow==4.2.1
prompt-toolkit==1.0.15    # via ipython, jupyter-console
protobuf==3.4.0           # via tensorflow, tensorflow-tensorboard
ptyprocess==0.5.2         # via pexpect, terminado
py==1.4.34                # via pytest
pycodestyle==2.3.1        # via flake8
pyflakes==1.5.0           # via flake8
pygments==2.2.0           # via ipython, jupyter-console, nbconvert, qtconsole
pyparsing==2.2.0          # via matplotlib
pytest-flake8==0.8.1
pytest-runner==2.12.1
pytest==3.2.2
python-dateutil==2.6.1    # via jupyter-client, matplotlib, pandas
pytz==2017.2              # via matplotlib, pandas
pyzmq==16.0.2             # via jupyter-client
qtconsole==4.3.1          # via jupyter
seaborn==0.8.1
simplegeneric==0.8.1      # via ipython
six==1.10.0
tensorflow-tensorboard==0.1.6  # via tensorflow
tensorflow-gpu==1.2.1     # added by keyonvafa
terminado==0.6            # via notebook
testpath==0.3.1           # via nbconvert
tornado==4.5.2            # via ipykernel, notebook, terminado
traitlets==4.3.2          # via ipykernel, ipython, ipywidgets, jupyter-client, jupyter-core, nbconvert, nbformat, notebook, qtconsole
wcwidth==0.1.7            # via prompt-toolkit
werkzeug==0.12.2          # via tensorflow-tensorboard
wheel==0.30.0             # via tensorflow, tensorflow-tensorboard
widgetsnbextension==3.0.2  # via ipywidgets
